Fix standalone compilation with gcc8.3
gcc 8.3 does not ship with the memory_resource header, but ships with
experimental/memory_resource

This header is required to use json without boost.

Since gcc 8.3 is what’s include in debian stable, it can be expected
that it will benefit many users to fallback on
experimental/memory_resource if possible.

Signed-off-by: Julien Blanc <julien.blanc@tgcm.eu>